+++ packages/data/src/types/VariantDescription.ts
@@ -1,3 +1,7 @@
+/**
+ * This is a basic description of a variant that will turned into a full VariantJSON object later
+ * on.
+ */
 export interface VariantDescription {
   name: string;
   suits: string[];
